requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
mysql-connector-python==8.0.27
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.4.1
psycopg2==2.9.3
pygame==2.1.2
Pygments==2.11.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.13.0
whitenoise==6.0.0
yarg==0.1.9